% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{feature_table}
\alias{feature_table}
\title{Wrapper function for \link[phyloseq]{otu_table}.}
\usage{
feature_table(object)
}
\arguments{
\item{object}{a feature table containing features at columns and samples at rows}
}
\value{
a \link[phyloseq]{phyloseq-class} feature table
}
\description{
Wrapper function for \link[phyloseq]{otu_table}.
}
\examples{
data(GlobalPatterns)
feature_table(test_phylo_object)
}
