% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phylo_constructor.R
\name{phylo_constructor}
\alias{phylo_constructor}
\title{Phyloseq constructor}
\usage{
phylo_constructor(microbiota_data, id_col, metadata)
}
\arguments{
\item{microbiota_data}{(Required). A data frame containing samples in rows and features with full taxonomy in cols.}

\item{id_col}{(Required). String for a column in the data frame to be used as sample id.}

\item{metadata}{(Optional). A data frame containing metadata for the experiment. Should contain the id_col.}
}
\value{
A \link[phyloseq]{phyloseq-class} object.
}
\description{
This function creates a \link[phyloseq]{phyloseq-class} object.
}
\examples{
phylo_constructor(test_data, id_col = "Infants_pdp_ID")
}
